Refactor caching implementation to use Flask-Caching
