+++ main.py
@@ -17,7 +17,7 @@ from blueprints.seo import seo
 from blueprints.stats import stats
 from blueprints.url_shortener import url_shortener
 from blueprints.redirector import url_redirector
-from services.cache import cache
+from services.flask_cache import cache
 from utils.mongo_utils import client
 
 app = Flask(__name__)
